rename logs_model.ts->logsModel.ts; update references
