+++ public/app/core/logsModel.ts
